+++ src/dynamicConfig.ts
@@ -196,7 +196,7 @@ export class DynamicConfig {
     public RefreshTmpGitFileListDuration: string = '10m';
     public AutoUpdateSearchTool: boolean = false;
     public CheckLanguageProcessIntervalMinutes = 15;
-    public OverwriteInConsistentCommonAliasByExtension = false;
+    public OverwriteInconsistentCommonAliasByExtension = false;
 
     private TmpToggleEnabledExtensionToValueMap = new Map<string, boolean>();
     private ProjectToGitIgnoreStatusMap = new Map<String, boolean>();
@@ -346,7 +346,7 @@ export class DynamicConfig {
         this.RefreshTmpGitFileListDuration = (getConfigValueOfActiveProject('refreshTmpGitFileListDuration', true) || '10m').replace(/\s+/g, '');
         this.AutoUpdateSearchTool = getConfigValueOfActiveProject('autoUpdateSearchTool') === 'true';
         this.CheckLanguageProcessIntervalMinutes = Math.max(5, Number(getConfigValueOfActiveProject('checkLanguageProcessIntervalMinutes') || '20'));
-        this.OverwriteInConsistentCommonAliasByExtension = getConfigValueOfActiveProject('overwriteInConsistentCommonAliasByExtension') === 'true';
+        this.OverwriteInconsistentCommonAliasByExtension = getConfigValueOfActiveProject('overwriteInconsistentCommonAliasByExtension') === 'true';
         SearchConfig.reload();
 
         this.ExcludeFoldersFromSettings.clear();